+++ lib/PhpParser/ConstExprEvaluationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace PhpParser;
+
+class ConstExprEvaluationException extends \Exception {}