+++ lib/PhpParser/Node/Expr/BinaryOp/Concat.php
@@ -6,4 +6,7 @@ use PhpParser\Node\Expr\BinaryOp;
 
 class Concat extends BinaryOp
 {
+    public function getOperatorSigil() : string {
+        return '.';
+    }
 }