+++ lib/PhpParser/Node/Expr/BinaryOp/Equal.php
@@ -6,4 +6,7 @@ use PhpParser\Node\Expr\BinaryOp;
 
 class Equal extends BinaryOp
 {
+    public function getOperatorSigil() : string {
+        return '==';
+    }
 }