+++ lib/PhpParser/Node/Expr/BinaryOp/Greater.php
@@ -6,4 +6,7 @@ use PhpParser\Node\Expr\BinaryOp;
 
 class Greater extends BinaryOp
 {
+    public function getOperatorSigil() : string {
+        return '>';
+    }
 }