+++ lib/PhpParser/Node/Expr/BinaryOp/Spaceship.php
@@ -6,4 +6,7 @@ use PhpParser\Node\Expr\BinaryOp;
 
 class Spaceship extends BinaryOp
 {
+    public function getOperatorSigil() : string {
+        return '<=>';
+    }
 }